+++ src/CB_PackingScalar.cpp
@@ -25,7 +25,7 @@
  *  @param [in] _VC 実際に送受信する仮想セル数
  *  @return 1次元インデクス
  */
-#define _IDX_SX(_I,_J,_K,_IS,_NJ,_VC) \
+#define _IDX_SI(_I,_J,_K,_IS,_NJ,_VC) \
 ( (_K+_VC) * _VC * (_NJ+2*_VC) \
 + (_J+_VC) * _VC \
 + (_I-_IS) \
@@ -40,7 +40,7 @@
  *  @param [in] _VC 実際に送受信する仮想セル数
  *  @return 1次元インデクス
  */
-#define _IDX_SY(_I,_J,_K,_NI,_JS,_VC) \
+#define _IDX_SJ(_I,_J,_K,_NI,_JS,_VC) \
 ( (_K+_VC) * (_NI+2*_VC) * _VC \
 + (_J-_JS) * (_NI+2*_VC) \
 + (_I+_VC) \
@@ -56,7 +56,7 @@
  *  @param [in] _VC 実際に送受信する仮想セル数
  *  @return 1次元インデクス
  */
-#define _IDX_SZ(_I,_J,_K,_NI,_NJ,_KS,_VC) \
+#define _IDX_SK(_I,_J,_K,_NI,_NJ,_KS,_VC) \
 ( (_K-_KS) * (_NI+2*_VC) * (_NJ+2*_VC) \
 + (_J+_VC) * (_NI+2*_VC) \
 + (_I+_VC) \
@@ -64,15 +64,15 @@
 
 
 /*
- * @brief pack send data for X direction
+ * @brief pack send data for I direction
  * @param [in]  array   source array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [out] sendm   send buffer of X- direction
- * @param [out] sendp   send buffer of X+ direction
- * @param [in]  nIDm    Rank number of X- direction
- * @param [in]  nIDp    Rank number of X+ direction
+ * @param [out] sendm   send buffer of I- direction
+ * @param [out] sendp   send buffer of I+ direction
+ * @param [in]  nIDm    Rank number of I- direction
+ * @param [in]  nIDp    Rank number of I+ direction
  */
-void SubDomain::pack_SX(const REAL_TYPE *array,
+void SubDomain::pack_SI(const REAL_TYPE *array,
                         const int vc_comm,
                         REAL_TYPE *sendm,
                         REAL_TYPE *sendp,
@@ -90,7 +90,7 @@ void SubDomain::pack_SX(const REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=0; i<vc_comm; i++ ){
-          sendm[_IDX_SX(i,j,k,0,jmax,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
+          sendm[_IDX_SI(i,j,k,0,jmax,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
         }
       }
     }
@@ -102,7 +102,7 @@ void SubDomain::pack_SX(const REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=imax-vc_comm; i<imax; i++ ){
-          sendp[_IDX_SX(i,j,k,imax-vc_comm,jmax,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
+          sendp[_IDX_SI(i,j,k,imax-vc_comm,jmax,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
         }
       }
     }
@@ -111,15 +111,15 @@ void SubDomain::pack_SX(const REAL_TYPE *array,
 
 
 /*
- * @brief unpack send data for X direction
+ * @brief unpack send data for I direction
  * @param [in,out]  array   dest array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [in]  recvm   recv buffer of X- direction
- * @param [in]  recvp   recv buffer of X+ direction
- * @param [in]  nIDm    Rank number of X- direction
- * @param [in]  nIDp    Rank number of X+ direction
+ * @param [in]  recvm   recv buffer of I- direction
+ * @param [in]  recvp   recv buffer of I+ direction
+ * @param [in]  nIDm    Rank number of I- direction
+ * @param [in]  nIDp    Rank number of I+ direction
  */
-void SubDomain::unpack_SX(REAL_TYPE *array,
+void SubDomain::unpack_SI(REAL_TYPE *array,
                           const int vc_comm,
                           const REAL_TYPE *recvm,
                           const REAL_TYPE *recvp,
@@ -137,7 +137,7 @@ void SubDomain::unpack_SX(REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=0-vc_comm; i<0; i++ ){
-          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvm[_IDX_SX(i,j,k,0-vc_comm,jmax,vc_comm)];
+          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvm[_IDX_SI(i,j,k,0-vc_comm,jmax,vc_comm)];
         }
       }
     }
@@ -149,7 +149,7 @@ void SubDomain::unpack_SX(REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=imax; i<imax+vc_comm; i++ ){
-          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvp[_IDX_SX(i,j,k,imax,jmax,vc_comm)];
+          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvp[_IDX_SI(i,j,k,imax,jmax,vc_comm)];
         }
       }
     }
@@ -157,15 +157,15 @@ void SubDomain::unpack_SX(REAL_TYPE *array,
 }
 
 /*
- * @brief pack send data for Y direction
+ * @brief pack send data for J direction
  * @param [in]  array   source array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [out] sendm   send buffer of Y- direction
- * @param [out] sendp   send buffer of Y+ direction
- * @param [in]  nIDm    Rank number of Y- direction
- * @param [in]  nIDp    Rank number of Y+ direction
+ * @param [out] sendm   send buffer of J- direction
+ * @param [out] sendp   send buffer of J+ direction
+ * @param [in]  nIDm    Rank number of J- direction
+ * @param [in]  nIDp    Rank number of J+ direction
  */
-void SubDomain::pack_SY(const REAL_TYPE *array,
+void SubDomain::pack_SJ(const REAL_TYPE *array,
                         const int vc_comm,
                         REAL_TYPE *sendm,
                         REAL_TYPE *sendp,
@@ -183,7 +183,7 @@ void SubDomain::pack_SY(const REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=0; j<vc_comm; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          sendm[_IDX_SY(i,j,k,imax,0,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
+          sendm[_IDX_SJ(i,j,k,imax,0,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
         }
       }
     }
@@ -195,7 +195,7 @@ void SubDomain::pack_SY(const REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=jmax-vc_comm; j<jmax; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          sendp[_IDX_SY(i,j,k,imax,jmax-vc_comm,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
+          sendp[_IDX_SJ(i,j,k,imax,jmax-vc_comm,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
         }
       }
     }
@@ -204,15 +204,15 @@ void SubDomain::pack_SY(const REAL_TYPE *array,
 
 
 /*
- * @brief unpack send data for Y direction
+ * @brief unpack send data for J direction
  * @param [in,out]  array   dest array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [in]  recvm   recv buffer of Y- direction
- * @param [in]  recvp   recv buffer of Y+ direction
- * @param [in]  nIDm    Rank number of Y- direction
- * @param [in]  nIDp    Rank number of Y+ direction
+ * @param [in]  recvm   recv buffer of J- direction
+ * @param [in]  recvp   recv buffer of J+ direction
+ * @param [in]  nIDm    Rank number of J- direction
+ * @param [in]  nIDp    Rank number of J+ direction
  */
-void SubDomain::unpack_SY(REAL_TYPE *array,
+void SubDomain::unpack_SJ(REAL_TYPE *array,
                           const int vc_comm,
                           const REAL_TYPE *recvm,
                           const REAL_TYPE *recvp,
@@ -230,7 +230,7 @@ void SubDomain::unpack_SY(REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=0-vc_comm; j<0; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvm[_IDX_SY(i,j,k,imax,0-vc_comm,vc_comm)];
+          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvm[_IDX_SJ(i,j,k,imax,0-vc_comm,vc_comm)];
         }
       }
     }
@@ -242,7 +242,7 @@ void SubDomain::unpack_SY(REAL_TYPE *array,
     for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
       for( int j=jmax; j<jmax+vc_comm; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvp[_IDX_SY(i,j,k,imax,jmax,vc_comm)];
+          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvp[_IDX_SJ(i,j,k,imax,jmax,vc_comm)];
         }
       }
     }
@@ -251,15 +251,15 @@ void SubDomain::unpack_SY(REAL_TYPE *array,
 
 
 /*
- * @brief pack send data for Z direction
+ * @brief pack send data for K direction
  * @param [in]  array   source array
  * @param [in]  vc_comm number of guide cell layer actually to be sent
- * @param [out] sendm   send buffer of Z- direction
- * @param [out] sendp   send buffer of Z+ direction
- * @param [in]  nIDm    Rank number of Z- direction
- * @param [in]  nIDp    Rank number of Z+ direction
+ * @param [out] sendm   send buffer of K- direction
+ * @param [out] sendp   send buffer of K+ direction
+ * @param [in]  nIDm    Rank number of K- direction
+ * @param [in]  nIDp    Rank number of K+ direction
  */
-void SubDomain::pack_SZ(const REAL_TYPE *array,
+void SubDomain::pack_SK(const REAL_TYPE *array,
                         const int vc_comm,
                         REAL_TYPE *sendm,
                         REAL_TYPE *sendp,
@@ -277,7 +277,7 @@ void SubDomain::pack_SZ(const REAL_TYPE *array,
     for( int k=0; k<vc_comm; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          sendm[_IDX_SZ(i,j,k,imax,jmax,0,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
+          sendm[_IDX_SK(i,j,k,imax,jmax,0,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
         }
       }
     }
@@ -289,7 +289,7 @@ void SubDomain::pack_SZ(const REAL_TYPE *array,
     for( int k=kmax-vc_comm; k<kmax; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          sendp[_IDX_SZ(i,j,k,imax,jmax,kmax-vc_comm,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
+          sendp[_IDX_SK(i,j,k,imax,jmax,kmax-vc_comm,vc_comm)] = array[_IDX_S3D(i,j,k,imax,jmax,vc)];
         }
       }
     }
@@ -298,15 +298,15 @@ void SubDomain::pack_SZ(const REAL_TYPE *array,
 
 
 /*
- * @brief unpack send data for Z direction
+ * @brief unpack send data for K direction
  * @param [in,out]  array   dest array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [in]  recvm   recv buffer of Z- direction
- * @param [in]  recvp   recv buffer of Z+ direction
- * @param [in]  nIDm    Rank number of Z- direction
- * @param [in]  nIDp    Rank number of Z+ direction
+ * @param [in]  recvm   recv buffer of K- direction
+ * @param [in]  recvp   recv buffer of K+ direction
+ * @param [in]  nIDm    Rank number of K- direction
+ * @param [in]  nIDp    Rank number of K+ direction
  */
-void SubDomain::unpack_SZ(REAL_TYPE *array,
+void SubDomain::unpack_SK(REAL_TYPE *array,
                           const int vc_comm,
                           const REAL_TYPE *recvm,
                           const REAL_TYPE *recvp,
@@ -324,7 +324,7 @@ void SubDomain::unpack_SZ(REAL_TYPE *array,
     for( int k=0-vc_comm; k<0; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvm[_IDX_SZ(i,j,k,imax,jmax,0-vc_comm,vc_comm)];
+          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvm[_IDX_SK(i,j,k,imax,jmax,0-vc_comm,vc_comm)];
         }
       }
     }
@@ -336,7 +336,7 @@ void SubDomain::unpack_SZ(REAL_TYPE *array,
     for( int k=kmax; k<kmax+vc_comm; k++ ){
       for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
         for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvp[_IDX_SZ(i,j,k,imax,jmax,kmax,vc_comm)];
+          array[_IDX_S3D(i,j,k,imax,jmax,vc)] = recvp[_IDX_SK(i,j,k,imax,jmax,kmax,vc_comm)];
         }
       }
     }