+++ src/CB_PackingVector.cpp
@@ -28,7 +28,7 @@
  *  @param [in] _VC 実際に送受信する仮想セル数
  *  @return 1次元インデクス
  */
-#define _IDX_VX(_I,_J,_K,_L,_IS,_NJ,_NK,_VC) \
+#define _IDX_VI(_I,_J,_K,_L,_IS,_NJ,_NK,_VC) \
 ( _L * _VC * (_NJ+2*_VC) * (_NK+2*_VC) \
 + (_K+_VC) * _VC * (_NJ+2*_VC) \
 + (_J+_VC) * _VC \
@@ -46,7 +46,7 @@
  *  @param [in] _VC 実際に送受信する仮想セル数
  *  @return 1次元インデクス
  */
-#define _IDX_VY(_I,_J,_K,_L,_NI,_JS,_NK,_VC) \
+#define _IDX_VJ(_I,_J,_K,_L,_NI,_JS,_NK,_VC) \
 ( _L * (_NI+2*_VC) * _VC * (_NK+2*_VC) \
 + (_K+_VC) * (_NI+2*_VC) * _VC \
 + (_J-_JS) * (_NI+2*_VC) \
@@ -64,7 +64,7 @@
  *  @param [in] _VC 実際に送受信する仮想セル数
  *  @return 1次元インデクス
  */
-#define _IDX_VZ(_I,_J,_K,_L,_NI,_NJ,_KS,_VC) \
+#define _IDX_VK(_I,_J,_K,_L,_NI,_NJ,_KS,_VC) \
 ( _L * (_NI+2*_VC) * (_NJ+2*_VC) * _VC \
 + (_K-_KS) * (_NI+2*_VC) * (_NJ+2*_VC) \
 + (_J+_VC) * (_NI+2*_VC) \
@@ -73,15 +73,15 @@
 
 
 /*
- * @brief pack send data for X direction
+ * @brief pack send data for I direction
  * @param [in]  array   source array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [out] sendm   send buffer of X- direction
- * @param [out] sendp   send buffer of X+ direction
- * @param [in]  nIDm    Rank number of X- direction
- * @param [in]  nIDp    Rank number of X+ direction
+ * @param [out] sendm   send buffer of I- direction
+ * @param [out] sendp   send buffer of I+ direction
+ * @param [in]  nIDm    Rank number of I- direction
+ * @param [in]  nIDp    Rank number of I+ direction
  */
-void SubDomain::pack_VX(const REAL_TYPE *array,
+void SubDomain::pack_VI(const REAL_TYPE *array,
                         const int vc_comm,
                         REAL_TYPE *sendm,
                         REAL_TYPE *sendp,
@@ -100,7 +100,7 @@ void SubDomain::pack_VX(const REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=0; i<vc_comm; i++ ){
-            sendm[_IDX_VX(i,j,k,l,0,jmax,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
+            sendm[_IDX_VI(i,j,k,l,0,jmax,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
           }
         }
       }
@@ -114,7 +114,7 @@ void SubDomain::pack_VX(const REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=imax-vc_comm; i<imax; i++ ){
-            sendp[_IDX_VX(i,j,k,l,imax-vc_comm,jmax,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
+            sendp[_IDX_VI(i,j,k,l,imax-vc_comm,jmax,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
           }
         }
       }
@@ -124,15 +124,15 @@ void SubDomain::pack_VX(const REAL_TYPE *array,
 
 
 /*
- * @brief unpack send data for X direction
+ * @brief unpack send data for I direction
  * @param [in,out]  array   dest array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [in]  recvm   recv buffer of X- direction
- * @param [in]  recvp   recv buffer of X+ direction
- * @param [in]  nIDm    Rank number of X- direction
- * @param [in]  nIDp    Rank number of X+ direction
+ * @param [in]  recvm   recv buffer of I- direction
+ * @param [in]  recvp   recv buffer of I+ direction
+ * @param [in]  nIDm    Rank number of I- direction
+ * @param [in]  nIDp    Rank number of I+ direction
  */
-void SubDomain::unpack_VX(REAL_TYPE *array,
+void SubDomain::unpack_VI(REAL_TYPE *array,
                           const int vc_comm,
                           const REAL_TYPE *recvm,
                           const REAL_TYPE *recvp,
@@ -151,7 +151,7 @@ void SubDomain::unpack_VX(REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=0-vc_comm; i<0; i++ ){
-            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvm[_IDX_VX(i,j,k,l,0-vc_comm,jmax,kmax,vc_comm)];
+            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvm[_IDX_VI(i,j,k,l,0-vc_comm,jmax,kmax,vc_comm)];
           }
         }
       }
@@ -165,7 +165,7 @@ void SubDomain::unpack_VX(REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=imax; i<imax+vc_comm; i++ ){
-            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvp[_IDX_VX(i,j,k,l,imax,jmax,kmax,vc_comm)];
+            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvp[_IDX_VI(i,j,k,l,imax,jmax,kmax,vc_comm)];
           }
         }
       }
@@ -174,15 +174,15 @@ void SubDomain::unpack_VX(REAL_TYPE *array,
 }
 
 /*
- * @brief pack send data for Y direction
+ * @brief pack send data for J direction
  * @param [in]  array   source array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [out] sendm   send buffer of Y- direction
- * @param [out] sendp   send buffer of Y+ direction
- * @param [in]  nIDm    Rank number of Y- direction
- * @param [in]  nIDp    Rank number of Y+ direction
+ * @param [out] sendm   send buffer of J- direction
+ * @param [out] sendp   send buffer of J+ direction
+ * @param [in]  nIDm    Rank number of J- direction
+ * @param [in]  nIDp    Rank number of J+ direction
  */
-void SubDomain::pack_VY(const REAL_TYPE *array,
+void SubDomain::pack_VJ(const REAL_TYPE *array,
                         const int vc_comm,
                         REAL_TYPE *sendm,
                         REAL_TYPE *sendp,
@@ -201,7 +201,7 @@ void SubDomain::pack_VY(const REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=0; j<vc_comm; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            sendm[_IDX_VY(i,j,k,l,imax,0,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
+            sendm[_IDX_VJ(i,j,k,l,imax,0,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
           }
         }
       }
@@ -215,7 +215,7 @@ void SubDomain::pack_VY(const REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=jmax-vc_comm; j<jmax; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            sendp[_IDX_VY(i,j,k,l,imax,jmax-vc_comm,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
+            sendp[_IDX_VJ(i,j,k,l,imax,jmax-vc_comm,kmax,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
           }
         }
       }
@@ -225,15 +225,15 @@ void SubDomain::pack_VY(const REAL_TYPE *array,
 
 
 /*
- * @brief unpack send data for Y direction
+ * @brief unpack send data for J direction
  * @param [in,out]  array   dest array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [in]  recvm   recv buffer of Y- direction
- * @param [in]  recvp   recv buffer of Y+ direction
- * @param [in]  nIDm    Rank number of Y- direction
- * @param [in]  nIDp    Rank number of Y+ direction
+ * @param [in]  recvm   recv buffer of J- direction
+ * @param [in]  recvp   recv buffer of J+ direction
+ * @param [in]  nIDm    Rank number of J- direction
+ * @param [in]  nIDp    Rank number of J+ direction
  */
-void SubDomain::unpack_VY(REAL_TYPE *array,
+void SubDomain::unpack_VJ(REAL_TYPE *array,
                           const int vc_comm,
                           const REAL_TYPE *recvm,
                           const REAL_TYPE *recvp,
@@ -252,7 +252,7 @@ void SubDomain::unpack_VY(REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=0-vc_comm; j<0; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvm[_IDX_VY(i,j,k,l,imax,0-vc_comm,kmax,vc_comm)];
+            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvm[_IDX_VJ(i,j,k,l,imax,0-vc_comm,kmax,vc_comm)];
           }
         }
       }
@@ -266,7 +266,7 @@ void SubDomain::unpack_VY(REAL_TYPE *array,
       for( int k=0-vc_comm; k<kmax+vc_comm; k++ ){
         for( int j=jmax; j<jmax+vc_comm; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvp[_IDX_VY(i,j,k,l,imax,jmax,kmax,vc_comm)];
+            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvp[_IDX_VJ(i,j,k,l,imax,jmax,kmax,vc_comm)];
           }
         }
       }
@@ -276,15 +276,15 @@ void SubDomain::unpack_VY(REAL_TYPE *array,
 
 
 /*
- * @brief pack send data for Z direction
+ * @brief pack send data for K direction
  * @param [in]  array   source array
  * @param [in]  vc_comm number of guide cell layer actually to be sent
- * @param [out] sendm   send buffer of Z- direction
- * @param [out] sendp   send buffer of Z+ direction
- * @param [in]  nIDm    Rank number of Z- direction
- * @param [in]  nIDp    Rank number of Z+ direction
+ * @param [out] sendm   send buffer of K- direction
+ * @param [out] sendp   send buffer of K+ direction
+ * @param [in]  nIDm    Rank number of K- direction
+ * @param [in]  nIDp    Rank number of K+ direction
  */
-void SubDomain::pack_VZ(const REAL_TYPE *array,
+void SubDomain::pack_VK(const REAL_TYPE *array,
                         const int vc_comm,
                         REAL_TYPE *sendm,
                         REAL_TYPE *sendp,
@@ -303,7 +303,7 @@ void SubDomain::pack_VZ(const REAL_TYPE *array,
       for( int k=0; k<vc_comm; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            sendm[_IDX_VZ(i,j,k,l,imax,jmax,0,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
+            sendm[_IDX_VK(i,j,k,l,imax,jmax,0,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
           }
         }
       }
@@ -317,7 +317,7 @@ void SubDomain::pack_VZ(const REAL_TYPE *array,
       for( int k=kmax-vc_comm; k<kmax; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            sendp[_IDX_VZ(i,j,k,l,imax,jmax,kmax-vc_comm,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
+            sendp[_IDX_VK(i,j,k,l,imax,jmax,kmax-vc_comm,vc_comm)] = array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)];
           }
         }
       }
@@ -327,15 +327,15 @@ void SubDomain::pack_VZ(const REAL_TYPE *array,
 
 
 /*
- * @brief unpack send data for Z direction
+ * @brief unpack send data for K direction
  * @param [in,out]  array   dest array
  * @param [in]  vc_comm number of guide cell layer to be sent
- * @param [in]  recvm   recv buffer of Z- direction
- * @param [in]  recvp   recv buffer of Z+ direction
- * @param [in]  nIDm    Rank number of Z- direction
- * @param [in]  nIDp    Rank number of Z+ direction
+ * @param [in]  recvm   recv buffer of K- direction
+ * @param [in]  recvp   recv buffer of K+ direction
+ * @param [in]  nIDm    Rank number of K- direction
+ * @param [in]  nIDp    Rank number of K+ direction
  */
-void SubDomain::unpack_VZ(REAL_TYPE *array,
+void SubDomain::unpack_VK(REAL_TYPE *array,
                           const int vc_comm,
                           const REAL_TYPE *recvm,
                           const REAL_TYPE *recvp,
@@ -354,7 +354,7 @@ void SubDomain::unpack_VZ(REAL_TYPE *array,
       for( int k=0-vc_comm; k<0; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvm[_IDX_VZ(i,j,k,l,imax,jmax,0-vc_comm,vc_comm)];
+            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvm[_IDX_VK(i,j,k,l,imax,jmax,0-vc_comm,vc_comm)];
           }
         }
       }
@@ -368,7 +368,7 @@ void SubDomain::unpack_VZ(REAL_TYPE *array,
       for( int k=kmax; k<kmax+vc_comm; k++ ){
         for( int j=0-vc_comm; j<jmax+vc_comm; j++ ){
           for( int i=0-vc_comm; i<imax+vc_comm; i++ ){
-            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvp[_IDX_VZ(i,j,k,l,imax,jmax,kmax,vc_comm)];
+            array[_IDX_V3D(i,j,k,l,imax,jmax,kmax,vc)] = recvp[_IDX_VK(i,j,k,l,imax,jmax,kmax,vc_comm)];
           }
         }
       }